chore: replace the existing block types
