+++ collab-document/src/importer/define.rs
@@ -1,91 +1,3 @@
-use crate::error::DocumentError;
-use std::fmt::Display;
-use std::str::FromStr;
-
-#[derive(Debug, PartialEq, Eq)]
-pub enum BlockType {
-  Page,
-  Paragraph,
-  Heading,
-  Quote,
-  TodoList,
-  NumberedList,
-  BulletedList,
-  Image,
-  LinkPreview,
-  Code,
-  MathEquation,
-  Divider,
-  Table,
-  TableCell,
-  Text,
-  Custom(String),
-}
-
-impl BlockType {
-  pub fn as_str(&self) -> &str {
-    match self {
-      BlockType::Page => "page",
-      BlockType::Paragraph => "paragraph",
-      BlockType::Heading => "heading",
-      BlockType::Quote => "quote",
-      BlockType::TodoList => "todo_list",
-      BlockType::NumberedList => "numbered_list",
-      BlockType::BulletedList => "bulleted_list",
-      BlockType::Image => "image",
-      BlockType::LinkPreview => "link_preview",
-      BlockType::Code => "code",
-      BlockType::MathEquation => "math_equation",
-      BlockType::Divider => "divider",
-      BlockType::Table => "table",
-      BlockType::TableCell => "table/cell",
-      BlockType::Text => "text",
-      BlockType::Custom(s) => s,
-    }
-  }
-
-  pub fn from_block_ty(s: &str) -> Self {
-    match s {
-      "page" => BlockType::Page,
-      "paragraph" => BlockType::Paragraph,
-      "heading" => BlockType::Heading,
-      "quote" => BlockType::Quote,
-      "todo_list" => BlockType::TodoList,
-      "numbered_list" => BlockType::NumberedList,
-      "bulleted_list" => BlockType::BulletedList,
-      "image" => BlockType::Image,
-      "link_preview" => BlockType::LinkPreview,
-      "code" => BlockType::Code,
-      "math_equation" => BlockType::MathEquation,
-      "divider" => BlockType::Divider,
-      "table" => BlockType::Table,
-      "table/cell" => BlockType::TableCell,
-      "text" => BlockType::Text,
-      _ => BlockType::Custom(s.to_string()),
-    }
-  }
-}
-impl FromStr for BlockType {
-  type Err = DocumentError;
-
-  fn from_str(s: &str) -> Result<Self, Self::Err> {
-    Ok(Self::from_block_ty(s))
-  }
-}
-
-// Implement AsRef<str> for ContentType
-impl AsRef<str> for BlockType {
-  fn as_ref(&self) -> &str {
-    self.as_str()
-  }
-}
-
-impl Display for BlockType {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    write!(f, "{}", self.as_str())
-  }
-}
-
 pub const IMAGE_EXTENSIONS: [&str; 6] = ["png", "jpg", "jpeg", "gif", "svg", "webp"];
 
 // Data Attribute Keys
