+++ collab-document/src/importer/md_importer.rs
@@ -1,4 +1,4 @@
-use crate::blocks::{Block, DocumentData, DocumentMeta};
+use crate::blocks::{Block, BlockType, DocumentData, DocumentMeta};
 use crate::document_data::generate_id;
 use crate::error::DocumentError;
 use crate::importer::define::*;
@@ -205,7 +205,7 @@ fn create_block(
     parent: parent_id.unwrap_or_default(),
     children: id.to_string(),
     external_id: Some(id.to_string()),
-    external_type: Some(BlockType::Text.to_string()),
+    external_type: Some(BlockType::Paragraph.to_string()),
   }
 }
 
@@ -336,12 +336,12 @@ fn create_table_cell_block(
 
   Block {
     id: id.to_string(),
-    ty: BlockType::TableCell.to_string(),
+    ty: BlockType::SimpleTableCell.to_string(),
     data: cell_data,
     parent: parent_id.to_string(),
     children: id.to_string(),
     external_id: Some(id.to_string()),
-    external_type: Some(BlockType::Text.to_string()),
+    external_type: Some(BlockType::Paragraph.to_string()),
   }
 }
